LayoutTests/http/tests/preload/dynamic_adding_preload-expected.txt

PASS internals.isPreloaded('../resources/dummy.js'); is true
PASS successfullyParsed is true

TEST COMPLETE
